[wip] refactor lib/ and work on test coverage
